refactor(common): one DSSE PAE, because two of them are two definitions of a signature
PAE is what a DSSE signature is actually taken over, so a second construction of
it is a second answer to what a signature covers. There were two, byte-identical,
in assay-evidence's mandate signing and assay-core's MCP signing. Byte-identical
is where a drift starts rather than a defence against one: nothing fails if one
side gains a space, renders a length differently, or handles a non-ASCII payload
type another way, until a signature made by one stops verifying on the other.

Found while preparing an AEE seal implementation, which would have written the
third copy. That is the point at which the standing rule applies rather than the
point at which it becomes urgent, so it is fixed before the third consumer exists.

assay_common::dsse::build_pae is the one construction; both call sites delegate.
It lives in assay-common because that is the crate both already depend on, and it
is std-gated for the same reason limits is: it allocates.

Tests pin the properties the two copies happened to share and neither stated:
lengths are byte lengths and not character counts, which is where a multi-byte
payload type would have split them; a payload carrying the separator byte cannot
restructure the encoding, which is what the length prefixes are for; and the
payload is appended verbatim including bytes that are not valid UTF-8.

CLAUDE.md described the assay-evidence -> assay-common edge as carrying only the
bounded-ingest primitive. It now carries two, so the doc states the test for
admitting a third: a mechanism whose second implementation would silently mean
something different. Vocabularies still stay domain-local.
