+++ crates/assay-common/src/lib.rs
@@ -10,6 +10,11 @@ pub mod exports;
 /// Bounded ingest primitive (ADR-043 §1). std-only: it is an `io::Read` adapter.
 pub mod limits;
 
+/// DSSE Pre-Authentication Encoding. std-only: it allocates. Shared because two
+/// implementations of PAE are two definitions of what a signature covers.
+#[cfg(feature = "std")]
+pub mod dsse;
+
 pub const EVENT_OPENAT: u32 = 1;
 pub const EVENT_CONNECT: u32 = 2;
 pub const EVENT_FORK: u32 = 3;
